requirements.txt
python-socketio[client]==5.0.3
spidev==3.5
Pillow==8.0.1
RPi.GPIO==0.7.0
evdev==1.3.0
